Added support for plugin banners
Usage: 
Add a new key named "banners" to the metadata file. It should be a JSON object with two string properties: "low" and "high". "low" must be a fully qualified URL pointing to a 772x250 image (PNG or JPG). "high" must point to a 1544x500 image. Only one of "low" or "high" is required. 

Example:
{
    "banners" : {
        "low" : "//example.com/assets/banner-772x250.png",
        "high" : "//example.com/assets/banner-1544x500.png"
    }
}

Banners are entirely optional.